Fix for objects library dataset downloads.
My best guess is there was a subtle bug in this code - because it was passing library dataset ids to an API endpoint that was expecting LDDA ids. LDDA ids would match LD ids in all the tests I guess because an LD never has its LDDA replaced I don't think.

Regardless - the API endpoint has changed so that now it does infact expect an LD id - but the parameter name has changed as well so I have fixed this here. More information on the Galaxy change can be found here - https://bitbucket.org/galaxy/galaxy-central/commits/c0c00d69cae3966a5c1e67abc3ba7aa67a735a64.

But this is definitely breaking - the new object library client will not work on older versions of Galaxy anymore and the older version of the client will not work with Galaxy after the October release. @simleo @nsoranzo is this acceptable - should we revisit the Galaxy API and make it more backward compatible to still allow LDDA IDs (I don't think it is worth it because of the bug mentioned above - but maybe I am wrong about the bug or maybe it is common to never replace a dataset so LD ids always match LDDA ids). Pinging @martenson on this as well.