Refactor make_post_request to compromise with comment from @nsoranzo.
Previously params=params was a no-op and I can respect that it looked silly - but I thought the symmetry of setting headers, params, and payload in both paths of the above if statement made the code more overtly correct.

Setting params or {} to a new variable continues to make this clear and results in a not no-op as requested by @nsoranzo.