Temporary fix for lack of floating ips
This is not the proper place for the fix - but in the current code,
there aren't many options. We can't do this on the cloudlaunch side
since cloudlaunch doesn't have an ec2 connection. We can't do this
properly on the bioblend side because the public IP is assigned after an
indeterminate waiting period. Luckily, none of that matter much since
most of this will be rewritten.