Remove wheel from ``requires`` in ``build-system`` section
See the following Note in https://setuptools.pypa.io/en/latest/userguide/quickstart.html#basic-use

"""
Historically this documentation has unnecessarily listed `wheel` in
the `requires` list, and many projects still do that. This is not
recommended. The backend automatically adds `wheel` dependency when
it is required, and listing it explicitly causes it to be
unnecessarily required for source distribution builds. You should
only include `wheel` in `requires` if you need to explicitly access
it during build time (e.g. if your project needs a `setup.py`
script that imports `wheel`).
"""